refactor scan modules
